Changed FO_Michelson to add uncertainty of the positions of the optics
This update adds the ability to add an error on the positioning of the optics (linear and angular). This is done though the modification of the Calc_geo function. In addition, the plot_geo function was added to display a representation of the optical system. This is usefull to ensure that the Calc_geo function works correctly.