Removed codes that did not work and modified FO_Michelson
Fourier_Optics_FP.py was removed as well as test_openphoton.py because they were not pertinent anymore since the 3 remaining codes can serve the purpose of evaluating the alignment precision for the 3 alignment setups.

The modifications to FO_Michelson were done to include the effects of the misalignement of the detector on the interference pattern. In essence a new function was added to do that (Detector_shift) and this function in used in all the plotting functions (except Plot_geo) to include the desired shift of the detector.